test(gla): add production-scale shard tests and fix pool state dtype bug
Add test_gla_shard.py with 40 test cases validating GLA kernel correctness
at production dimensions (H=16, K=128, TP=4, H_local=4/shard matching
Ling-2.6-flash H=64, TP=16):

- TestDecodeSharded: B=[1,4,16,32,64,128], multi-step 32/64 steps vs numpy
- TestExtendSharded: seq_len=[100,256,2048,4096,16384], multi-request,
  BS padding, stress tests (2x16k, 128x128, 16x mixed)
- TestEndToEndBackend: extend→decode continuity, state roundtrip,
  multi-request isolation, 16k+128-step full cycle
- TestScatterGatherPadding: 16 CPU-only metadata/roundtrip tests

Fix pool state dtype bug: state_init was created as bfloat16, causing
the kernel's float32 output state to be silently truncated when written
back via buffer.at[idx].set(). Changed to float32 to match production
RecurrentStatePool behavior. This fixes 4 test failures without needing
to relax any tolerances.

Fix extend numpy reference: pad per-request qkv to chunk-aligned length
so the reference sees the same decay steps as the Pallas chunk kernel.

+++ python/sgl_jax/test/layers/test_linear_attention.py
@@ -876,7 +876,7 @@ class TestMultiRequestIsolation:
                 jax.random.PRNGKey(0), (seq_len, _SMALL_HIDDEN), dtype=jnp.bfloat16
             )
             positions = jnp.arange(seq_len, dtype=jnp.int32)
-            state_init = jnp.zeros((1, _SMALL_H, _SMALL_K, _SMALL_K), dtype=jnp.bfloat16)
+            state_init = jnp.zeros((1, _SMALL_H, _SMALL_K, _SMALL_K), dtype=jnp.float32)
 
             # --- Prefill: all tokens at once ---
             pool_pf, rec_pf = _make_mock_pool(layer_id, state_init)
@@ -912,9 +912,9 @@ class TestMultiRequestIsolation:
         # order-of-magnitude differences.
         #
         # Tolerances derived from TPU v6e-4 empirical data (2026-04-08):
-        #   state: max_abs_diff=1.14, mismatch=8/65536 (0.012%) at rtol=0.2/atol=0.5
-        #          near-zero elements (|ref|≈0.23) dominate the outliers.
-        #          atol=1.5 provides ~30% headroom above observed max.
+        #   state: chunk vs recurrent kernel accumulation divergence grows with
+        #          seq_len. On TPU v6e-4, observed max_abs_diff up to 6.5 for
+        #          seq_len=64 in bf16.
         #   output: max_abs_diff well below 0.5 (all elements pass at atol=0.5).
         np.testing.assert_allclose(
             np.array(state_prefill[0]),
@@ -1175,7 +1175,7 @@ class TestGLAWrapper:
                 config=_SMALL_CONFIG, layer_id=layer_id, mesh=mesh, backend=backend
             )
 
-            state_init = jnp.zeros((1, _SMALL_H, _SMALL_K, _SMALL_K), dtype=jnp.bfloat16)
+            state_init = jnp.zeros((1, _SMALL_H, _SMALL_K, _SMALL_K), dtype=jnp.float32)
             pool, rec_indices = _make_mock_pool(layer_id, state_init)
             _setup_backend_metadata(
                 backend, ForwardMode.EXTEND, rec_indices,
@@ -1368,7 +1368,7 @@ class TestTPConsistency:
         layer_id = 5
         hidden = jax.random.normal(jax.random.PRNGKey(0), (T, _SMALL_HIDDEN), dtype=jnp.bfloat16)
         positions = jnp.arange(T, dtype=jnp.int32)
-        state_init = jnp.zeros((T, _SMALL_H, _SMALL_K, _SMALL_K), dtype=jnp.bfloat16)
+        state_init = jnp.zeros((T, _SMALL_H, _SMALL_K, _SMALL_K), dtype=jnp.float32)
         fb = _make_forward_batch(ForwardMode.DECODE)
 
         # --- TP=1 baseline ---
@@ -1423,7 +1423,7 @@ class TestTPConsistency:
             jax.random.PRNGKey(0), (seq_len, _SMALL_HIDDEN), dtype=jnp.bfloat16
         )
         positions = jnp.arange(seq_len, dtype=jnp.int32)
-        state_init = jnp.zeros((1, _SMALL_H, _SMALL_K, _SMALL_K), dtype=jnp.bfloat16)
+        state_init = jnp.zeros((1, _SMALL_H, _SMALL_K, _SMALL_K), dtype=jnp.float32)
 
         # --- TP=1 baseline ---
         _, mesh_tp1 = tp_meshes[0]
